bot/requirements.txt
instagrapi==2.0.0
python-telegram-bot==20.3
flask==2.3.3
schedule==1.2.0
requests==2.31.0
